requirements/base.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5
autopep8==1.5.5
certifi==2020.12.5
cloudinary==1.26.0
colorama==0.4.4
distlib==0.3.1
Django==3.1.7
django-tinymce==3.3.0
djangorestframework==3.12.4
filelock==3.0.12
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mysql-connector-python==8.0.25
Pillow==8.4.0
pipenv==2020.11.15
pycodestyle==2.7.0
pylint==2.7.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
virtualenv==20.4.2
virtualenv-clone==0.5.4
wrapt==1.12.1
